requirements.txt
git+http://github.com/mitsuhiko/flask.git
git+http://github.com/mitsuhiko/flask-sqlachemy.git
Flask-Assets
cssmin
psycopg2
